# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  models.MaintenanceRecord:
    properties:
      asset_code:
        type: string
      created_at:
        type: string
      description:
        type: string
      equipment_id:
        type: string
      equipment_name:
        type: string
      is_resolved:
        type: boolean
      lid:
        type: string
      reporter_type:
        description: '''public'', ''staff'', ''system'''
        type: string
      resolve_note:
        type: string
    type: object
  models.MaintenanceRequest:
    properties:
      description:
        maxLength: 500
        minLength: 1
        type: string
      equipment_id:
        type: string
    required:
    - description
    - equipment_id
    type: object
  models.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  models.ResolveMaintenanceRequest:
    properties:
      lid:
        type: string
      resolve_note:
        maxLength: 500
        minLength: 1
        type: string
    required:
    - lid
    - resolve_note
    type: object
...
info:
  contact: {}
  description: 這是一個整合 PostgreSQL 與 Gin 的設備管理系統。
  title: 設備管理系統 API
  version: "1.0"
paths:
...
  /api/private/maintenance-records:
    get:
      consumes:
      - application/json
      description: 取得所有維修紀錄，支援狀態篩選與分頁 (僅限管理員與維修人員)
      parameters:
      - description: 是否顯示已解決 (true/false)
        in: query
        name: resolved
        type: string
      - description: 每頁筆數 (預設 50，最大 100)
        in: query
        name: limit
        type: integer
      - description: 偏移量 (預設 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: 符合條件的總筆數
              type: integer
          schema:
            items:
              $ref: '#/definitions/models.MaintenanceRecord'
            type: array
        "400":
          description: 查詢參數錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 取得維修紀錄列表
      tags:
      - private
  /api/private/maintenance-records/resolve:
    patch:
      consumes:
      - application/json
      description: 將維修紀錄標記為已解決，並將設備狀態恢復為 normal。會驗證紀錄是否已經解決避免重複操作。
      parameters:
      - description: 維修完成請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ResolveMaintenanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功修復
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 紀錄不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 紀錄已被解決
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 標記維修完成
      tags:
      - private
...
  /api/public/report:
    post:
      consumes:
      - application/json
      description: 建立一個新的報修紀錄，並將設備狀態更新為 faulty。由 server 端固定 reporter_type 為 public
        並限制請求大小。
      parameters:
      - description: 報修資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MaintenanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功提交報修紀錄
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 設備不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 設備已有未處理紀錄
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 提交報修紀錄
      tags:
      - public
securityDefinitions:
  BearerAuth:
    description: 請輸入 "Bearer <Your_JWT_Token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"